+++ src/tui/tui.go
@@ -132,6 +132,7 @@ const (
 	Jump
 	JumpCancel
 	ClickHeader
+	Multi
 )
 
 func (t EventType) AsEvent() Event {